Consolidate annotation code
