+++ fonduer/__init__.py
@@ -29,8 +29,11 @@ from fonduer.meta import Meta
 from fonduer.parser import Parser
 from fonduer.parser.models import Document, Figure, Sentence, Table
 from fonduer.parser.preprocessors import HTMLDocPreprocessor
-from fonduer.supervision.annotations import load_gold_labels
-from fonduer.supervision.async_annotations import FeatureAnnotator, LabelAnnotator
+from fonduer.supervision.annotations import (
+    FeatureAnnotator,
+    LabelAnnotator,
+    load_gold_labels,
+)
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
